# Speedup esp32 flash